Move ServletHandlerFactory to core
Needed to move to access ThreadNameFilter.
