+++ dropwizard-core/src/main/java/com/codahale/dropwizard/server/ServletHandlerFactory.java
@@ -1,5 +1,7 @@
-package com.codahale.dropwizard.jetty;
+package com.codahale.dropwizard.server;
 
+import com.codahale.dropwizard.jetty.MutableServletContextHandler;
+import com.codahale.dropwizard.server.DefaultServletHandlerFactory;
 import com.codahale.metrics.MetricRegistry;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
