+++ dropwizard-core/src/test/java/com/codahale/dropwizard/server/ContextServletHandlerFactoryTest.java
@@ -1,4 +1,4 @@
-package com.codahale.dropwizard.jetty;
+package com.codahale.dropwizard.server;
 
 /** TODO: Document */
 public class ContextServletHandlerFactoryTest {
