+++ dropwizard-core/src/test/resources/yaml/handler.yml
